backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.1

# Database
alembic==1.13.1
aiosqlite==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0

# Documentation
python-multipart==0.0.9
email-validator==2.1.0.post1

# Development tools
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Optional dependencies
matplotlib==3.8.2  # For budget visualization
numpy==1.26.3     # For calculations 
gunicorn
